requirements.txt
discord.py
d20
